Split the inspector into its own module, and fix the bugbot + review findings
Extract the inspection code into fused_render/templates/pdf_studio/inspector.py
(828 lines), leaving pdf.py at 1535 instead of 2324. pdf.py loads it by path
via importlib, not by name: the built-in executor runs template helpers with
sys.path untouched, so a sibling import would not resolve — and under the
subprocess engine the script's own directory IS on sys.path, where a module
named `inspector` could shadow something. Coupling is one-way and small: pdf.py
resolves the working-copy path and page selection, then calls report(path, idx)
or markdown(path, idx).

Cursor bugbot findings:
- _rust_engine caught only ImportError, so any failure inside pdf_inspector
  aborted the whole inspect. The optional cross-check can no longer take the
  report down: absent it reports the install command, present-but-failing it
  reports the error, and the pymupdf verdict stands either way.
- pikepdf.NameTree raised on a malformed /Names/JavaScript. The tree is now
  walked by hand, type-gated, so a broken script tree costs its own findings
  instead of the whole scan.
- The bare-URL linkifier rewrote URLs inside code spans and indented code
  lines, turning code into nested links.
- clearOpenDoc did not drop the report, so removing the last library entry left
  the previous document's inspection on screen.
- _font_report opened a second document handle and never closed it; it now
  reuses the one already open.

Second review pass:
- _wrap_link replaced the first substring match, which hit "here" inside
  "Adherence", the anchor of an earlier link on the same line, and words inside
  another link's URL. It now matches whole words and skips existing links and
  code spans.
- A damaged object graph no longer costs the entire report: the scan degrades
  to an "unreadable" verdict, which is itself the finding, while pages, fonts
  and classification still render.
- A failed scan left the drawer showing "Scanning…" with no way back.
